refactor: 统一 Groovy 脚本相关服务到 script/service 目录
- 移除重复的 services 和 utils 目录，统一使用 script/service
- 合并 TitleConfigModal 到 node-title-config-modal
- 将变量选择器样式从 less 迁移到 scss
- 添加样式规范到 CLAUDE.md
