+++ frontend/packages/flow-pc/flow-pc-design/src/components/design-editor/script/service/index.ts
@@ -1,3 +1,4 @@
 export { GroovySyntaxConverter, groovySyntaxConverter } from './groovy-syntax-converter';
 export { GroovyVariableService, groovyVariableService } from './groovy-variable-service';
+export { TitleSyntaxConverter } from './title-syntax-converter';
 export { ScriptAdapterManager, scriptAdapterManager } from './adapters';
