--- frontend/packages/flow-pc/flow-pc-design/src/services/index.ts
@@ -1,1 +1,0 @@
-export * from './groovy-variable-service';
