+++ frontend/packages/flow-pc/flow-pc-design/src/type.d.ts
@@ -2,3 +2,4 @@ declare module '*.svg'
 declare module '*.png'
 declare module '*.jpg'
 declare module '*.module.less'
+declare module '*.module.scss'
