--- frontend/packages/flow-pc/flow-pc-design/src/utils/index.ts
@@ -1,1 +1,0 @@
-export * from './title-syntax-converter';
